add some comments about registration and remove a not-needed class
