+++ src/PKSim.UI.Starter/PKSimStarterPresenterRegister.cs
@@ -1,9 +1,6 @@
 ﻿using OSPSuite.Utility.Container;
-using PKSim.Core;
 using PKSim.Presentation;
 using PKSim.Presentation.Presenters.Individuals;
-using PKSim.Presentation.UICommands;
-using PKSim.UI.UICommands;
 
 namespace PKSim.UI.Starter
 {
@@ -24,20 +21,6 @@ namespace PKSim.UI.Starter
       protected override void RegisterMainViewPresenters(IContainer container)
       {
          
-      }
-   }
-
-   public class PKSimStarterUserInterfaceRegister : Register
-   {
-      public override void RegisterInContainer(IContainer container)
-      {
-         container.AddScanner(scan =>
-         {
-            scan.AssemblyContainingType<UserInterfaceRegister>();
-            scan.WithConvention<PKSimStarterRegistrationConvention>();
-         });
-         container.Register<OSPSuite.UI.Services.IToolTipCreator, IToolTipCreator, ToolTipCreator>(LifeStyle.Transient);
-         container.Register<IExitCommand, ExitCommand>();
       }
    }
 }
